clv_acquisition_targeting/requirements.txt
numpy>=1.24,<3
pandas>=2.0,<3
scikit-learn>=1.3,<2
matplotlib>=3.7,<4
seaborn>=0.13,<1
PyYAML>=6.0,<7
requests>=2.31,<3
lxml>=5.0,<6
jupyter>=1.0,<2
nbconvert>=7,<8
ipykernel>=6,<7
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1

# Optional (see config.yaml model_type: xgboost):
# xgboost>=2.0,<4
